+++ client/data/image.json
@@ -233,6 +233,12 @@
     },
     {
       "id": 78
+    },
+    {
+      "id": 79
+    },
+    {
+      "id": 80
     }
   ]
 }
