--- client/src/pages/UserEdit.js
@@ -1,7 +1,0 @@
-// import { UserEditBox } from "../components/UserEdit/UserEdit";
-import Diary2 from "./Diary/Diary2/Diary2";
-function UserEdit() {
-  return <Diary2 />;
-}
-
-export default UserEdit;
